@@ -1,13 +1,13 @@
-from sphinx.util.docutils import SphinxDirective
-from sphinx.locale import _
-from docutils import nodes
-from sphinx import addnodes
-
-from collections import OrderedDict, namedtuple
 import importlib
 import inspect
 import os
 import re
+from collections import OrderedDict, namedtuple
+
+from docutils import nodes
+from sphinx import addnodes
+from sphinx.locale import _
+from sphinx.util.docutils import SphinxDirective
 
 
 class attributetable(nodes.General, nodes.Element):
